Add more experimental move outs
