+++ ttnn/api/ttnn/tensor/tensor_utils.hpp
@@ -11,6 +11,7 @@
 #include <tt-metalium/program_descriptors.hpp>
 
 // Exports symbols
+#include <tt-metalium/experimental/distributed_tensor/distributed_tensor_apis.hpp>
 #include <tt-metalium/experimental/tensor/tensor_apis.hpp>
 
 namespace ttnn {
